Culture feature and application settings support
