+++ src/Radical.Windows/Presentation/Boot/Features/CurrentPrincipal.cs
@@ -6,7 +6,7 @@ namespace Radical.Windows.Presentation.Boot.Features
 {
     class CurrentPrincipal : IFeature
     {
-        public void Setup(IServiceProvider serviceProvider)
+        public void Setup(IServiceProvider serviceProvider, ApplicationSettings applicationSettings)
         {
             Thread.CurrentPrincipal = new WindowsPrincipal(WindowsIdentity.GetCurrent());
         }
