+++ src/Radical.Windows/Presentation/Boot/Features/UIComposition.cs
@@ -6,7 +6,7 @@ namespace Radical.Windows.Presentation.Boot.Features
 {
     class UIComposition : IFeature
     {
-        public void Setup(IServiceProvider serviceProvider)
+        public void Setup(IServiceProvider serviceProvider, ApplicationSettings applicationSettings)
         {
             RegionService.CurrentService = serviceProvider.GetService<IRegionService>();
             RegionService.Conventions = serviceProvider.GetService<IConventionsHandler>();
